Add a .gitignore
I add the .gitignore and told it to ignore some of the more common
files that editors likes to create, i.e. backup stuff. Keep them
locally, don't push them.
Also swaps the first and last name of Andreas, which was the only
name in the .tex that was formated like that. But that's a minor
thing.