Add the Scripts folder and the pylint_test_suite.py script.
The Scripts folder will contain useful scripts used in the project,
but not directly a part of the code that we will present.
That is the case with the pylint_test_suite.py script. This is
intended to be used as a pre-commit hook to check any and all python
code we will push to the repo. If the code doesn't pass a certain
coding standard, then the commit will be aborted.